Fix: Correct project structure and improve login test
This commit addresses several issues in the project setup and test suite.

- The project structure has been corrected to follow Maven conventions by moving all test-related source files from `src/main/java` to `src/test/java`.
- The `pom.xml` has been updated to set the correct `test` scope for the TestNG dependency and to use a compatible Java version (21).
- The `ClientLogin` test has been refactored into a single, atomic test method in `ClientLoginTest.java`. This new test now includes a proper assertion to verify that the login is successful by checking for the presence of the user avatar on the homepage.
- The WebDriver initialization in `Base.java` has been configured to run Chrome in headless mode to support execution in non-graphical environments.
